Delay the preCollection remove after upload to workaround sporadic timing issue. Fixes #324
Though the call to Collection.update should be blocking (server side) according to metoer API:
https://docs.meteor.com/api/collections.html#Mongo-Collection-update
In fact the update sometimes is not completed before the remove kicks in, which in turn triggers the preCollection.obeserve("remove") code. This leads to the situation that the oberver code sees an unfinished upload while the upload actually finished correctly.