v1.8.1
 - Exact code splitting, client part of the library reduced for 35%
 - Atmosphere and NPM packages update
 - Compatibility with meteor@1.5.1
 - Fix for #458 - Safari can't handle simultaneous WebSockets
connection, and fails to send large data chunks. It's fixed in this
release with detecting if browser is Safari with `ddp` used as
transport. Note is added to FAQ, Should be fixed in Safari 11. Thanks
to @derwok for bringing it in, and @IDCOLL for original issue
 - Fix resumable uploads, now in 90% upload should survive server
reboot and in 98% connection interruption
 - Better file stream management with file descriptor caching - should
reduce RAM and I/O